Remove library copies and generate symlinks during build
The .so versions of the libraries are just copies of the versioned
libraries with a different name. This is just bloat in the repository,
so create and cleanup symbolic links during the build. The libraries
need to exist with the .so filename since that's the actual SONAME.

Creating the symlinks in the rules and ignoring them in git makes it
easier on the next source import since the links don't have to be
manually recreated.